+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG for FlatCAM Evo beta
 - Paint Plugin: fixed issues caused by the latest changes in the Shapely module
 - NCC Plugin: some changes in the method used by the Tcl Command
 - in Preferences, added a control in the General -> GUI Settings to control the font size for the entire application; require an app restart
+- Sub Tool: fixed an issue due of an incorrect parameter for a QTimer start
 
 31.05.2023